split up files more logically
